Conflicts:
	frappe/__init__.py
	frappe/core/doctype/user/user.py
	frappe/core/doctype/user_permission/user_permission.json
	frappe/desk/doctype/todo/test_todo.js
	frappe/desk/doctype/todo/todo.json
	frappe/public/css/page.css
	frappe/public/js/frappe/form/control.js
	frappe/public/js/frappe/form/grid.js
	frappe/public/js/frappe/form/templates/grid_body.html
	frappe/public/js/frappe/form/templates/print_layout.html
	frappe/public/js/frappe/ui/base_list.js
	frappe/public/js/frappe/ui/page.js
	frappe/public/js/frappe/views/reports/query_report.js
	frappe/templates/styles/classic.css
	frappe/templates/styles/modern.css
	frappe/utils/backups.py
	frappe/utils/pdf.py
	frappe/utils/print_format.py
	frappe/utils/verified_command.py
	frappe/website/doctype/blog_post/templates/blog_post_row.html
	frappe/website/doctype/contact_us_settings/contact_us_settings.json
	frappe/website/doctype/contact_us_settings/test_contact_us_settings.js
	frappe/website/doctype/web_form/web_form.json
	frappe/website/doctype/web_page/test_web_page.js
	frappe/website/doctype/web_page/web_page.json
	frappe/website/render.py
	frappe/www/printview.py

+++ frappe/utils/backups.py
@@ -60,10 +60,15 @@ class BackupGenerator:
 	def set_backup_file_name(self):
 		todays_date = datetime.today().strftime("%d%m%Y_%H%M%S")
 
-		#Generate a name using today's date with time
-		for_db = todays_date + "_database.sql.gz"
-		for_public_files = todays_date + "_files.tar"
-		for_private_files = todays_date + "_private_files.tar"
+
+		todays_date = now_datetime().strftime('%Y%m%d_%H%M%S')
+		site = frappe.local.site or frappe.generate_hash(length=8)
+		site = site.replace('.', '_')
+
+		#Generate a random name using today's date and a 8 digit random number
+		for_db = todays_date + "-" + site + "-database.sql.gz"
+		for_public_files = todays_date + "-" + site + "-files.tar"
+		for_private_files = todays_date + "-" + site + "-private-files.tar"
 		backup_path = get_backup_path()
 
 		if not self.backup_path_db:
@@ -177,11 +182,13 @@ def delete_temp_backups(older_than=24):
 	"""
 		Cleans up the backup_link_path directory by deleting files older than 24 hours
 	"""
-	file_list = os.listdir(get_backup_path())
-	for this_file in file_list:
-		this_file_path = os.path.join(get_backup_path(), this_file)
-		if is_file_old(this_file_path, older_than):
-			os.remove(this_file_path)
+	backup_path = get_backup_path()
+	if os.path.exists(backup_path):
+		file_list = os.listdir(get_backup_path())
+		for this_file in file_list:
+			this_file_path = os.path.join(get_backup_path(), this_file)
+			if is_file_old(this_file_path, older_than):
+				os.remove(this_file_path)
 
 def is_file_old(db_file_name, older_than=24):
 		"""
